changes so far

+++ Http/controllers/restaurant/table/table.destroy.php
@@ -16,9 +16,22 @@ $table = $db->query('select * from restaurant_table where tableid = :id', [
 
 authorize($table['resID'] === $userid);
 
+
+
+if($table['status']==1){
 $db->query('delete from restaurant_table where tableid = :id', [
     'id' => $_POST['tableid']
 ]);
 
 header('location: /tables');
 exit();
+
+
+}
+
+
+header('location: /tables');
+exit();
+
+
+
